pyproject.toml
[tool.poetry]
name = "mastodon-trend-analysis"
version = "0.1.0"
description = "This project aims to analyse trends on Mastodon Fediverse"
authors = ["Artur Khaialiev <a12020032@unet.univie.ac.at>"]

[tool.poetry.dependencies]
python = "^3.9"
sanic = "^22.12.0"
alembic = "^1.9.1"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.5"
sanic-session = "^0.8.0"
aiohttp = "^3.8.3"
mypy = "^0.991"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}
sqlalchemy-utils = "^0.39.0"
greenlet = "^2.0.1"
python-dotenv = "^0.21.0"
sanic-ext = "^22.12.0"
orjson = "^3.8.4"
aioschedule = "^0.5.2"
sanic-jwt = "^1.8.0"
pydantic-sqlalchemy = "^0.0.9"
mastodon-py = "^1.8.0"
pytz = "^2022.7.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
